requirements.txt
streamlit
pydantic
pillow
sqlite3
